requirements_runpod.txt
# Core dependencies for RunPod deployment
torch==2.1.0
torchvision==0.16.0
transformers>=4.30.0
accelerate>=0.20.0
tokenizers>=0.13.0
sentencepiece
peft>=0.5.0
bitsandbytes>=0.41.0

# Video processing
decord==0.6.0
opencv-python-headless
ffmpeg-python
moviepy==1.0.3

# Audio processing
soundfile==0.12.1

# RunPod and utilities
runpod
python-dotenv
numpy==1.24.4
pillow
requests